Refactor logic for 'createGrid' function
This commit changes the 'createGrid' function to create 16 rows within
the grid and fill each row with 16 squares. This was done to ensure
each square is cleanly placed within a row.

To ensure the divs resemble actual squares that conform to the space of
row, and the entire container itself, the following CSS classes were created
that the function dynamically applies to the newly created divs:
    - 'grid-row' allows the row to grow or shrink based on the number
       of other rows and squares within them.
    - 'grid-square' allows the individual squares to grow or shrink
       based on the height of the row and the number of squares within
       the row.

The height and width CSS properties were applied to the #grid-container
rule to give the container defined boundaries that the child elements
must conform to.

These changes were also made to ensure the function is easily updated
in the future, where it will take a number argument to create the grid
instead of a defined 16x16.

Various border properties were added for testing purposes.