Add functions 'createGridRow' & 'createGridSquare'
This commit adds the functions 'createGridRow' & 'createGridSquare' to
main.js, and calls them in the 'createGrid' function. The change ensures
the 'createGrid' function easier to read by following the Single
Responsability Principle.

The 'createGrid' function is now only responsible for bringing the
different parts of the grid together (thus creating the grid), while the
'createGridRow' and 'createGridSquare' functions are now responsible
for creating the individual parts themselves.