Create RGB sliders with HTML and style with CSS
This commit only adds the basic HTML structure for the RGB sliders and
adds basic styling with CSS. Their functionality will be added at a
later date using JavaScript.

In the future these RGB sliders will change the color to be drawn onto
the grid so that the user is not only restricted to black.