Refactor logic for clearing grid on button click
This commit improves the efficiency of the function. In the previous
iteration, it would delete the grid, and rebuild it to clear all
markings made by the user. While this did clear the grid, it did so with
extra steps that were not necessary. The logic now removes all instances
of 'background-color' styling but keeps the grid intact.

In addition, all function calls and variable declarations made outside
a function declaration are contained within the 'main' function.